Update tests to include new functionality
