+++ Minio.Functional.Tests/FunctionalTest.cs
@@ -45,280 +45,8 @@ using Minio.Helper;
 namespace Minio.Functional.Tests;
 
 [SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Keep private const lowercase")]
-public static class FunctionalTest
+public static partial class FunctionalTest
 {
-    private const int KB = 1024;
-    private const int MB = 1024 * 1024;
-    private const int GB = 1024 * 1024 * 1024;
-
-    private const string dataFile1B = "datafile-1-b";
-
-    private const string dataFile10KB = "datafile-10-kB";
-    private const string dataFile6MB = "datafile-6-MB";
-
-    private const string makeBucketSignature =
-        "Task MakeBucketAsync(string bucketName, string location = 'us-east-1', CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string listBucketsSignature =
-        "Task<ListAllMyBucketsResult> ListBucketsAsync(CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string bucketExistsSignature =
-        "Task<bool> BucketExistsAsync(string bucketName, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeBucketSignature =
-        "Task RemoveBucketAsync(string bucketName, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string listObjectsSignature =
-        "IObservable<Item> ListObjectsAsync(string bucketName, string prefix = null, bool recursive = false, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string putObjectSignature =
-        "Task PutObjectAsync(PutObjectArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getObjectSignature =
-        "Task GetObjectAsync(GetObjectArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string listIncompleteUploadsSignature =
-        "IObservable<Upload> ListIncompleteUploads(ListIncompleteUploads args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string listenBucketNotificationsSignature =
-        "IObservable<MinioNotificationRaw> ListenBucketNotificationsAsync(ListenBucketNotificationsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string listenNotificationsSignature =
-        "IObservable<MinioNotificationRaw> ListenNotifications(ListenBucketNotificationsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string copyObjectSignature =
-        "Task<CopyObjectResult> CopyObjectAsync(CopyObjectArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string statObjectSignature =
-        "Task<ObjectStat> StatObjectAsync(StatObjectArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeObjectSignature1 =
-        "Task RemoveObjectAsync(RemoveObjectArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeObjectSignature2 =
-        "Task<IObservable<DeleteError>> RemoveObjectsAsync(RemoveObjectsArgs, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeIncompleteUploadSignature =
-        "Task RemoveIncompleteUploadAsync(RemoveIncompleteUploadArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string presignedPutObjectSignature =
-        "Task<string> PresignedPutObjectAsync(PresignedPutObjectArgs args)";
-
-    private const string presignedGetObjectSignature =
-        "Task<string> PresignedGetObjectAsync(PresignedGetObjectArgs args)";
-
-    private const string presignedPostPolicySignature =
-        "Task<Dictionary<string, string>> PresignedPostPolicyAsync(PresignedPostPolicyArgs args)";
-
-    private const string getBucketPolicySignature =
-        "Task<string> GetPolicyAsync(GetPolicyArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setBucketPolicySignature =
-        "Task SetPolicyAsync(SetPolicyArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getBucketNotificationSignature =
-        "Task<BucketNotification> GetBucketNotificationAsync(GetBucketNotificationsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setBucketNotificationSignature =
-        "Task SetBucketNotificationAsync(SetBucketNotificationsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeAllBucketsNotificationSignature =
-        "Task RemoveAllBucketNotificationsAsync(RemoveAllBucketNotifications args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setBucketEncryptionSignature =
-        "Task SetBucketEncryptionAsync(SetBucketEncryptionArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getBucketEncryptionSignature =
-        "Task<ServerSideEncryptionConfiguration> GetBucketEncryptionAsync(GetBucketEncryptionArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeBucketEncryptionSignature =
-        "Task RemoveBucketEncryptionAsync(RemoveBucketEncryptionArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string selectObjectSignature =
-        "Task<SelectResponseStream> SelectObjectContentAsync(SelectObjectContentArgs args,CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setObjectLegalHoldSignature =
-        "Task SetObjectLegalHoldAsync(SetObjectLegalHoldArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getObjectLegalHoldSignature =
-        "Task<bool> GetObjectLegalHoldAsync(GetObjectLegalHoldArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setObjectLockConfigurationSignature =
-        "Task SetObjectLockConfigurationAsync(SetObjectLockConfigurationArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getObjectLockConfigurationSignature =
-        "Task<ObjectLockConfiguration> GetObjectLockConfigurationAsync(GetObjectLockConfigurationArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string deleteObjectLockConfigurationSignature =
-        "Task RemoveObjectLockConfigurationAsync(GetObjectLockConfigurationArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getBucketTagsSignature =
-        "Task<Tagging> GetBucketTagsAsync(GetBucketTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setBucketTagsSignature =
-        "Task SetBucketTagsAsync(SetBucketTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string deleteBucketTagsSignature =
-        "Task RemoveBucketTagsAsync(RemoveBucketTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setVersioningSignature =
-        "Task SetVersioningAsync(SetVersioningArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getVersioningSignature =
-        "Task<VersioningConfiguration> GetVersioningAsync(GetVersioningArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string removeVersioningSignature =
-        "Task RemoveBucketTagsAsync(RemoveBucketTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getObjectTagsSignature =
-        "Task<Tagging> GetObjectTagsAsync(GetObjectTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setObjectTagsSignature =
-        "Task SetObjectTagsAsync(SetObjectTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string deleteObjectTagsSignature =
-        "Task RemoveObjectTagsAsync(RemoveObjectTagsArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setObjectRetentionSignature =
-        "Task SetObjectRetentionAsync(SetObjectRetentionArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getObjectRetentionSignature =
-        "Task<ObjectRetentionConfiguration> GetObjectRetentionAsync(GetObjectRetentionArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string clearObjectRetentionSignature =
-        "Task ClearObjectRetentionAsync(ClearObjectRetentionArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string getBucketLifecycleSignature =
-        "Task<LifecycleConfiguration> GetBucketLifecycleAsync(GetBucketLifecycleArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string setBucketLifecycleSignature =
-        "Task SetBucketLifecycleAsync(SetBucketLifecycleArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private const string deleteBucketLifecycleSignature =
-        "Task RemoveBucketLifecycleAsync(RemoveBucketLifecycleArgs args, CancellationToken cancellationToken = default(CancellationToken))";
-
-    private static readonly Random rnd = new();
-
-    private static readonly RandomStreamGenerator rsg = new(100 * MB);
-
-    private static string Bash(string cmd)
-    {
-        var Replacements = new Dictionary<string, string>
-            (StringComparer.Ordinal)
-            {
-                { "$", "\\$" },
-                { "(", "\\(" },
-                { ")", "\\)" },
-                { "{", "\\{" },
-                { "}", "\\}" },
-                { "[", "\\[" },
-                { "]", "\\]" },
-                { "@", "\\@" },
-                { "%", "\\%" },
-                { "&", "\\&" },
-                { "#", "\\#" },
-                { "+", "\\+" }
-            };
-
-        foreach (var toReplace in Replacements.Keys)
-            cmd = cmd.Replace(toReplace, Replacements[toReplace], StringComparison.Ordinal);
-        var cmdNoReturn = cmd + " >/dev/null 2>&1";
-
-        using var process = new Process
-        {
-            StartInfo = new ProcessStartInfo
-            {
-                FileName = "/bin/bash",
-                Arguments = $"-c \"{cmdNoReturn}\"",
-                UseShellExecute = false,
-                RedirectStandardOutput = true,
-                CreateNoWindow = true
-            }
-        };
-
-        _ = process.Start();
-        var result = process.StandardOutput.ReadLine();
-        process.WaitForExit();
-
-        return result;
-    }
-
-    // Create a file of given size from random byte array or optionally create a symbolic link
-    // to the dataFileName residing in MINT_DATA_DIR
-    private static string CreateFile(int size, string dataFileName = null)
-    {
-        var fileName = GetRandomName();
-
-        if (!IsMintEnv())
-        {
-            var data = new byte[size];
-            rnd.NextBytes(data);
-
-            File.WriteAllBytes(fileName, data);
-            return GetFilePath(fileName);
-        }
-
-        return GetFilePath(dataFileName);
-    }
-
-    public static string GetRandomObjectName(int length = 5)
-    {
-        // Server side does not allow the following characters in object names
-        // '-', '_', '.', '/', '*'
-#if NET6_0_OR_GREATER
-        var characters = "abcd+%$#@&{}[]()";
-#else
-        var characters = "abcdefgh+%$#@&";
-#endif
-        var result = new StringBuilder(length);
-
-        for (var i = 0; i < length; i++) result.Append(characters[rnd.Next(characters.Length)]);
-        return result.ToString();
-    }
-
-    // Generate a random string
-    public static string GetRandomName(int length = 5)
-    {
-        var characters = "0123456789abcdefghijklmnopqrstuvwxyz";
-        if (length > 50) length = 50;
-
-        var result = new StringBuilder(length);
-        for (var i = 0; i < length; i++) _ = result.Append(characters[rnd.Next(characters.Length)]);
-
-        return "minio-dotnet-example-" + result;
-    }
-
-    internal static void GenerateRandom500MB_File(string fileName)
-    {
-        using var fs = new FileStream(fileName, FileMode.Create, FileAccess.Write, FileShare.None, 4096, true);
-        var fileSize = 500L * 1024 * 1024;
-        var segments = fileSize / 10000;
-        var last_seg = fileSize % 10000;
-        using var br = new BinaryWriter(fs);
-
-        for (long i = 0; i < segments; i++)
-            br.Write(new byte[10000]);
-
-        br.Write(new byte[last_seg]);
-        br.Close();
-    }
-
-    // Return true if running in Mint mode
-    public static bool IsMintEnv()
-    {
-        return !string.IsNullOrEmpty(Environment.GetEnvironmentVariable("MINT_DATA_DIR"));
-    }
-
-    // Get full path of file
-    public static string GetFilePath(string fileName)
-    {
-        var dataDir = Environment.GetEnvironmentVariable("MINT_DATA_DIR");
-        if (!string.IsNullOrEmpty(dataDir)) return $"{dataDir}/{fileName}";
-
-        var path = Directory.GetCurrentDirectory();
-        return $"{path}/{fileName}";
-    }
-
     internal static Task RunCoreTests(IMinioClient minioClient)
     {
         ConcurrentBag<Task> coreTestsTasks = new()
@@ -340,6 +68,8 @@ public static class FunctionalTest
             // Test Presigned Get/Put operations
             PresignedGetObject_Test1(minioClient),
             PresignedPutObject_Test1(minioClient),
+            GetPresignedUrl_Get_Test1(minioClient),
+            GetPresignedUrl_Put_Delete_Test1(minioClient),
 
             // Test incomplete uploads
             ListIncompleteUpload_Test1(minioClient),
@@ -1319,6 +1049,13 @@ public static class FunctionalTest
         using var filestream = new FileStream(filePath, FileMode.Open, FileAccess.Read);
         using var stream = new StreamContent(filestream);
         await minio.WrapperPutAsync(url, stream).ConfigureAwait(false);
+    }
+
+    internal static async Task DeleteObjectAsync(IMinioClient minio, string url)
+    {
+        using var response = await minio.WrapperDeleteAsync(url).ConfigureAwait(false);
+        if (string.IsNullOrEmpty(Convert.ToString(response.Content)) || HttpStatusCode.OK != response.StatusCode)
+            throw new InvalidOperationException("Unable to delete via presigned URL" + nameof(response.Content));
     }
 
     internal static async Task PresignedPostPolicy_Test1(IMinioClient minio)
@@ -5940,6 +5677,367 @@ public static class FunctionalTest
         {
             new MintLogger("PresignedPutObject_Test2", presignedPutObjectSignature,
                 "Tests whether PresignedPutObject url retrieves object from bucket when invalid expiry is set.",
+                TestStatus.FAIL, DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
+            throw;
+        }
+        finally
+        {
+            await TearDown(minio, bucketName).ConfigureAwait(false);
+        }
+    }
+
+    #endregion
+
+    #region Get Presigned Url
+
+    internal static async Task GetPresignedUrl_Get_Test1(IMinioClient minio)
+    {
+        var startTime = DateTime.Now;
+        var bucketName = GetRandomName(15);
+        var objectName = GetRandomObjectName(10);
+        var expiresInt = 1000;
+        var downloadFile = GetRandomObjectName(10);
+
+        var args = new Dictionary<string, string>
+            (StringComparer.Ordinal)
+            {
+                { "bucketName", bucketName },
+                { "objectName", objectName },
+                { "expiresInt", expiresInt.ToString(CultureInfo.InvariantCulture) }
+            };
+        try
+        {
+            await Setup_Test(minio, bucketName).ConfigureAwait(false);
+            using (var filestream = rsg.GenerateStreamFromSeed(1 * KB))
+            {
+                var putObjectArgs = new PutObjectArgs()
+                    .WithBucket(bucketName)
+                    .WithObject(objectName)
+                    .WithStreamData(filestream)
+                    .WithObjectSize(filestream.Length);
+
+                _ = await minio.PutObjectAsync(putObjectArgs).ConfigureAwait(false);
+            }
+
+            var statObjectArgs = new StatObjectArgs()
+                .WithBucket(bucketName)
+                .WithObject(objectName);
+            var stats = await minio.StatObjectAsync(statObjectArgs).ConfigureAwait(false);
+
+            var httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Get;
+            var preArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(expiresInt);
+            var presigned_url = await minio.GetPresignedUrlAsync(preArgs).ConfigureAwait(false);
+
+            await DownloadObjectAsync(minio, presigned_url, downloadFile).ConfigureAwait(false);
+            var writtenInfo = new FileInfo(downloadFile);
+            var file_read_size = writtenInfo.Length;
+            // Compare the size of the file downloaded using the generated
+            // presigned_url (expected value) and the actual object size on the server
+            Assert.AreEqual(file_read_size, stats.Size);
+            new MintLogger("GetPresignedUrl_Get_Test1", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket", TestStatus.PASS,
+                DateTime.Now - startTime, args: args).Log();
+        }
+        catch (Exception ex)
+        {
+            new MintLogger("GetPresignedUrl_Get_Test1", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket", TestStatus.FAIL,
+                DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
+            throw;
+        }
+        finally
+        {
+            File.Delete(downloadFile);
+            await TearDown(minio, bucketName).ConfigureAwait(false);
+        }
+    }
+
+    internal static async Task GetPresignedUrl_Get_Test2(IMinioClient minio)
+    {
+        var startTime = DateTime.Now;
+        var bucketName = GetRandomName(15);
+        var objectName = GetRandomObjectName(10);
+        var expiresInt = 0;
+        var args = new Dictionary<string, string>
+            (StringComparer.Ordinal)
+            {
+                { "bucketName", bucketName },
+                { "objectName", objectName },
+                { "expiresInt", expiresInt.ToString(CultureInfo.InvariantCulture) }
+            };
+        try
+        {
+            await Setup_Test(minio, bucketName).ConfigureAwait(false);
+            using (var filestream = rsg.GenerateStreamFromSeed(1 * KB))
+            {
+                var putObjectArgs = new PutObjectArgs()
+                    .WithBucket(bucketName)
+                    .WithObject(objectName)
+                    .WithStreamData(filestream)
+                    .WithObjectSize(filestream.Length);
+
+                _ = await minio.PutObjectAsync(putObjectArgs).ConfigureAwait(false);
+            }
+
+            var statObjectArgs = new StatObjectArgs()
+                .WithBucket(bucketName)
+                .WithObject(objectName);
+            var stats = await minio.StatObjectAsync(statObjectArgs).ConfigureAwait(false);
+            var httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Get;
+            var preArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(0);
+            var presigned_url = await minio.GetPresignedUrlAsync(preArgs).ConfigureAwait(false);
+            throw new InvalidOperationException(
+                "GetPresignedUrlAsync expected to throw an InvalidExpiryRangeException.");
+        }
+        catch (InvalidExpiryRangeException)
+        {
+            new MintLogger("GetPresignedUrl_Get_Test2", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when invalid expiry is set.",
+                TestStatus.PASS, DateTime.Now - startTime, args: args).Log();
+        }
+        catch (InvalidOperationException ex)
+        {
+            new MintLogger("GetPresignedUrl_Get_Test2", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when invalid expiry is set.",
+                TestStatus.FAIL, DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
+            throw;
+        }
+        catch (Exception ex)
+        {
+            new MintLogger("GetPresignedUrl_Get_Test2", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when invalid expiry is set.",
+                TestStatus.FAIL, DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
+            throw;
+        }
+        finally
+        {
+            await TearDown(minio, bucketName).ConfigureAwait(false);
+        }
+    }
+
+    internal static async Task GetPresignedUrl_Get_Test3(IMinioClient minio)
+    {
+        var startTime = DateTime.Now;
+        var bucketName = GetRandomName(15);
+        var objectName = GetRandomObjectName(10);
+        var expiresInt = 1000;
+        var reqDate = DateTime.UtcNow.AddSeconds(-50);
+        var downloadFile = GetRandomObjectName(10);
+        var args = new Dictionary<string, string>
+            (StringComparer.Ordinal)
+            {
+                { "bucketName", bucketName },
+                { "objectName", objectName },
+                { "expiresInt", expiresInt.ToString(CultureInfo.InvariantCulture) },
+                {
+                    "reqParams",
+                    "response-content-type:application/json,response-content-disposition:attachment;filename=  MyDoc u m  e   nt.json ;"
+                },
+                { "reqDate", reqDate.ToString(CultureInfo.InvariantCulture) }
+            };
+        try
+        {
+            await Setup_Test(minio, bucketName).ConfigureAwait(false);
+            using (var filestream = rsg.GenerateStreamFromSeed(1 * KB))
+            {
+                var putObjectArgs = new PutObjectArgs()
+                    .WithBucket(bucketName)
+                    .WithObject(objectName)
+                    .WithStreamData(filestream)
+                    .WithObjectSize(filestream.Length);
+
+                _ = await minio.PutObjectAsync(putObjectArgs).ConfigureAwait(false);
+            }
+
+            var statObjectArgs = new StatObjectArgs()
+                .WithBucket(bucketName)
+                .WithObject(objectName);
+            var stats = await minio.StatObjectAsync(statObjectArgs).ConfigureAwait(false);
+            var reqParams = new Dictionary<string, string>
+                (StringComparer.Ordinal)
+                {
+                    ["response-content-type"] = "application/json",
+                    ["response-content-disposition"] = "attachment;filename=  MyDoc u m  e   nt.json ;"
+                };
+            var httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Get;
+            var preArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(1000)
+                .WithHeaders(reqParams)
+                .WithRequestDate(reqDate);
+            var presigned_url = await minio.GetPresignedUrlAsync(preArgs).ConfigureAwait(false);
+
+            using var response = await minio.WrapperGetAsync(presigned_url).ConfigureAwait(false);
+            if (response.StatusCode != HttpStatusCode.OK ||
+                string.IsNullOrEmpty(Convert.ToString(response.Content, CultureInfo.InvariantCulture)))
+                throw new InvalidOperationException("Unable to download via presigned URL " + nameof(response.Content));
+
+            Assert.IsTrue(response.Content.Headers.GetValues("Content-Type")
+                .Contains(reqParams["response-content-type"], StringComparer.Ordinal));
+            Assert.IsTrue(response.Content.Headers.GetValues("Content-Disposition")
+                .Contains(reqParams["response-content-disposition"], StringComparer.Ordinal));
+            Assert.IsTrue(response.Content.Headers.GetValues("Content-Length")
+                .Contains(stats.Size.ToString(CultureInfo.InvariantCulture), StringComparer.Ordinal));
+
+            using (var fs = new FileStream(downloadFile, FileMode.CreateNew))
+            {
+                await response.Content.CopyToAsync(fs).ConfigureAwait(false);
+            }
+
+            var writtenInfo = new FileInfo(downloadFile);
+            var file_read_size = writtenInfo.Length;
+
+            // Compare the size of the file downloaded with the generated
+            // presigned_url (expected) and the actual object size on the server
+            Assert.AreEqual(file_read_size, stats.Size);
+            new MintLogger("GetPresignedUrl_Get_Test3", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when override response headers sent",
+                TestStatus.PASS, DateTime.Now - startTime, args: args).Log();
+        }
+        catch (Exception ex)
+        {
+            new MintLogger("GetPresignedUrl_Get_Test3", presignedGetObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when override response headers sent",
+                TestStatus.FAIL, DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
+            throw;
+        }
+        finally
+        {
+            File.Delete(downloadFile);
+            await TearDown(minio, bucketName).ConfigureAwait(false);
+        }
+    }
+
+    internal static async Task GetPresignedUrl_Put_Delete_Test1(IMinioClient minio)
+    {
+        var startTime = DateTime.Now;
+        var bucketName = GetRandomName(15);
+        var objectName = GetRandomObjectName(10);
+        var expiresInt = 1000;
+        var fileName = CreateFile(10 * KB, dataFile10KB);
+
+        var args = new Dictionary<string, string>
+            (StringComparer.Ordinal)
+            {
+                { "bucketName", bucketName },
+                { "objectName", objectName },
+                { "expiresInt", expiresInt.ToString(CultureInfo.InvariantCulture) }
+            };
+        try
+        {
+            await Setup_Test(minio, bucketName).ConfigureAwait(false);
+            // Upload with presigned url
+            var httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Put;
+            var presignedPutObjectArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(1000);
+            var presigned_url = await minio.GetPresignedUrlAsync(presignedPutObjectArgs).ConfigureAwait(false);
+            await UploadObjectAsync(minio, presigned_url, fileName).ConfigureAwait(false);
+            // Get stats for object from server
+            var statObjectArgs = new StatObjectArgs()
+                .WithBucket(bucketName)
+                .WithObject(objectName);
+            var stats = await minio.StatObjectAsync(statObjectArgs).ConfigureAwait(false);
+            // Compare with file used for upload
+            var writtenInfo = new FileInfo(fileName);
+            var file_written_size = writtenInfo.Length;
+            Assert.AreEqual(file_written_size, stats.Size);
+            new MintLogger("GetPresignedUrl_Put_Delete_Test1", presignedPutObjectSignature,
+                "Tests whether GetPresignedUrl url uploads object to bucket", TestStatus.PASS,
+                DateTime.Now - startTime, args: args).Log();
+
+            // Delete with presigned url
+            httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Delete;
+            presignedPutObjectArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(1000);
+            presigned_url = await minio.GetPresignedUrlAsync(presignedPutObjectArgs).ConfigureAwait(false);
+            await DeleteObjectAsync(minio, presigned_url).ConfigureAwait(false);
+        }
+        catch (Exception ex)
+        {
+            new MintLogger("GetPresignedUrl_Put_Delete_Test1", presignedPutObjectSignature,
+                "Tests whether GetPresignedUrl url uploads object to bucket", TestStatus.FAIL,
+                DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
+            throw;
+        }
+        finally
+        {
+            await TearDown(minio, bucketName).ConfigureAwait(false);
+            if (!IsMintEnv()) File.Delete(fileName);
+        }
+    }
+
+    internal static async Task GetPresignedUrl_Put_Delete_Test2(IMinioClient minio)
+    {
+        var startTime = DateTime.Now;
+        var bucketName = GetRandomName(15);
+        var objectName = GetRandomObjectName(10);
+        var expiresInt = 0;
+
+        var args = new Dictionary<string, string>
+            (StringComparer.Ordinal)
+            {
+                { "bucketName", bucketName },
+                { "objectName", objectName },
+                { "expiresInt", expiresInt.ToString(CultureInfo.InvariantCulture) }
+            };
+        try
+        {
+            await Setup_Test(minio, bucketName).ConfigureAwait(false);
+            using (var filestream = rsg.GenerateStreamFromSeed(1 * KB))
+            {
+                var putObjectArgs = new PutObjectArgs()
+                    .WithBucket(bucketName)
+                    .WithObject(objectName)
+                    .WithStreamData(filestream)
+                    .WithObjectSize(filestream.Length);
+
+                _ = await minio.PutObjectAsync(putObjectArgs).ConfigureAwait(false);
+            }
+
+            var statObjectArgs = new StatObjectArgs()
+                .WithBucket(bucketName)
+                .WithObject(objectName);
+            var stats = await minio.StatObjectAsync(statObjectArgs).ConfigureAwait(false);
+            var httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Put;
+            var presignedPutObjectArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(0);
+            var presigned_url = await minio.GetPresignedUrlAsync(presignedPutObjectArgs).ConfigureAwait(false);
+            new MintLogger("GetPresignedUrl_Put_Delete_Test2", presignedPutObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when invalid expiry is set.",
+                TestStatus.PASS, DateTime.Now - startTime, args: args).Log();
+
+            // Delete with presigned url
+            httpMethod = GetPresignedUrlArgs.PresignedUrlHttpMethod.Delete;
+            presignedPutObjectArgs = new GetPresignedUrlArgs(httpMethod)
+                .WithBucket(bucketName)
+                .WithObject(objectName)
+                .WithExpiry(1000);
+            presigned_url = await minio.GetPresignedUrlAsync(presignedPutObjectArgs).ConfigureAwait(false);
+            await DeleteObjectAsync(minio, presigned_url).ConfigureAwait(false);
+        }
+        catch (InvalidExpiryRangeException)
+        {
+            new MintLogger("GetPresignedUrl_Put_Delete_Test2", presignedPutObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when invalid expiry is set.",
+                TestStatus.PASS, DateTime.Now - startTime, args: args).Log();
+        }
+        catch (Exception ex)
+        {
+            new MintLogger("GetPresignedUrl_Put_Delete_Test2", presignedPutObjectSignature,
+                "Tests whether GetPresignedUrl url retrieves object from bucket when invalid expiry is set.",
                 TestStatus.FAIL, DateTime.Now - startTime, ex.Message, ex.ToString(), args: args).Log();
             throw;
         }
