+++ Minio.Functional.Tests/Program.cs
@@ -212,6 +212,11 @@ internal static class Program
         functionalTestTasks.Add(FunctionalTest.PresignedPutObject_Test1(minioClient));
         functionalTestTasks.Add(FunctionalTest.PresignedPutObject_Test2(minioClient));
         // FunctionalTest.PresignedPostPolicy_Test1(minioClient).Wait();
+        functionalTestTasks.Add(FunctionalTest.GetPresignedUrl_Get_Test1(minioClient));
+        functionalTestTasks.Add(FunctionalTest.GetPresignedUrl_Get_Test2(minioClient));
+        functionalTestTasks.Add(FunctionalTest.GetPresignedUrl_Get_Test3(minioClient));
+        functionalTestTasks.Add(FunctionalTest.GetPresignedUrl_Put_Delete_Test1(minioClient));
+        functionalTestTasks.Add(FunctionalTest.GetPresignedUrl_Put_Delete_Test2(minioClient));
 
         // Test incomplete uploads
         functionalTestTasks.Add(FunctionalTest.ListIncompleteUpload_Test1(minioClient));
